int + test

+++ src/main/java/com/example/encheres/bo/ArticleVendu.java
@@ -11,8 +11,8 @@ public class ArticleVendu {
 	private String description;
 	private LocalDate dateDebutEnchere;
 	private LocalDate dateFinEnchere;
-	private float prixInitial;
-	private float prixVente;
+	private int prixInitial;
+	private int prixVente;
 
 	private Categorie categorie;
 
@@ -23,7 +23,7 @@ public class ArticleVendu {
 
 
 public ArticleVendu(int noArticle, String nomArticle, String description, LocalDate dateDebutEnchere,
-			LocalDate dateFinEnchere, float prixInitial, float prixVente, Categorie categorie, Utilisateur vendeur, Utilisateur acheteur,
+			LocalDate dateFinEnchere, int prixInitial, int prixVente, Categorie categorie, Utilisateur vendeur, Utilisateur acheteur,
 			List<Enchere> encheres) {
 		super();
 		this.noArticle = noArticle;
@@ -94,7 +94,7 @@ public float getPrixInitial() {
 	return prixInitial;
 }
 
-public void setPrixInitial(float prixInitial) {
+public void setPrixInitial(int prixInitial) {
 	this.prixInitial = prixInitial;
 }
 
@@ -102,7 +102,7 @@ public float getPrixVente() {
 	return prixVente;
 }
 
-public void setPrixVente(float prixVente) {
+public void setPrixVente(int prixVente) {
 	this.prixVente = prixVente;
 }
